move alert component to components
