+++ src/pages/Estimates/EstimatesGen.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { Button } from '@material-ui/core';
 import { PDF } from '../../components/Common/PDF';
-import { SaveAlert } from '../../components/Common/Alerts';
+import { SaveAlert } from '../../components/Alerts';
 import EstimatesTemplate from './EstimatesTemplate';
 
 const EstimatesGen = props => {
